src/backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.3
Flask-MQTT==1.2.1
python-dotenv==1.0.1
psycopg2==2.9.9
gunicorn==20.1.0